Language Identification
In natural language processing, language identification is the problem
of determining which natural language given content is in. Computational
approaches to this problem view it as a special case of text
categorization, solved with various statistical methods. The data set is
lines of movie subtitles from 21 different languages. The task here is
to train a model based on training data set, learn the hyper parameters
on validation set and identify the language of subtitles in unknown
languages. The model implemented for this task is Naïve Bayes model.
Each line in training data set is considered as a single document. All
documents are represented as Bag of n-grams.